security implemented + the DEVIN issue resolved
The Devin issue recap:
- change container name from snake case to camel case
"current_user" to "currentUser"
-  to access the session attributes in thymeleaf we use
${session.currentUser.name} or ${session.currentUser.email}
- setting the attribute we instantiate HttpSession object on the parameters level and use session.setAttribute method instead of a model
- @SessionAttributes(value = {"currentUser"}) must be used on class level preferably on a controller that will be called after login since this will rely on Principal object so a user should be already logged in, therefore i use an if statement to check if a principal is null or not.
- if you need to access the session attributes we can do that with two methods:
the first: use HttpSession session on parameter level and getAttribute method.
the second: use @SessionAttribute("currentUser") on any method available on a controller to access it.

note it's imported that to init the session with values before beginning to use it in the app otherwise we get exceptions like field not found.